fix: rename from global suppressions to bulk suppressions
