--- build-tests/eslint-global-suppressions-test/.eslint-global-suppressions.json
@@ -1,10 +1,0 @@
-{
-  "suppressions": [
-    {
-      "file": "src/test-lab.ts",
-      "scope": ".ClassDeclaration.MethodDefinition.ClassExpression.MethodDefinition.ObjectExpression.FunctionExpression",
-      "target": ".AbsurdClass.absurdClassMethod.AbsurdClass2.constructor.absurdObject.absurdObjectMethod",
-      "rule": "@typescript-eslint/no-empty-function"
-    }
-  ]
-}
