+++ eslint/eslint-config/patch/eslint-bulk-suppressions.js
@@ -1,4 +1,4 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-require('@rushstack/eslint-patch/lib/eslint-global-suppressions');
+require('@rushstack/eslint-patch/eslint-bulk-suppressions');
