+++ eslint/eslint-patch/eslint-bulk-suppressions.js
@@ -1,0 +1,1 @@
+require('./lib/eslint-bulk-suppressions');
