+++ eslint/eslint-patch/src/eslint-bulk-suppressions/ast-node-type-guards.ts
